cinqo de mayooo

+++ original_libft/ft_strlcat.c
@@ -6,21 +6,21 @@
 /*   By: ssergiu <ssergiu@student.42heilbronn.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/30 19:47:55 by ssergiu           #+#    #+#             */
-/*   Updated: 2022/05/03 17:56:03 by ssergiu          ###   ########.fr       */
+/*   Updated: 2022/05/05 16:17:11 by ssergiu          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	int		i;
+	size_t	i;
 	size_t	dstlen;
 
 	dstlen = ft_strlen(dst);
 	i = 0;
 	if (dstsize <= dstlen)
 		return (ft_strlen(src) + dstsize);
-	while (i < (int)(dstsize - 1 - dstlen) && src[i])
+	while (i < (dstsize - 1 - dstlen) && src[i])
 	{
 		dst[i + dstlen] = src[i];
 		i++;
